+++ src/tools/textEncryptor/textEncryptor.js
